api/requirements.txt
aiohttp==3.8.3
aiomysql==0.1.1
aiosignal==1.3.1
aiosmtplib==3.0.1
anyio==3.6.2
apimatic-core==0.2.8
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
async-timeout==4.0.2
attrs==22.2.0
CacheControl==0.12.14
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.2
deprecation==2.1.0
dnspython==2.3.0
email-validator==2.0.0.post2
enum34==1.1.10
exceptiongroup==1.1.0
fastapi==0.89.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jsonpickle==3.0.2
jsonpointer==2.4
mangum==0.17.0
MarkupSafe==2.1.2
msgpack==1.0.7
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
squareup==34.0.1.20240118
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==2.2.0
uvicorn==0.20.0
yarl==1.8.2
pyOpenSSL
